apps/api/requirements/development.txt
-r base.txt

# Development-only dependencies
django-debug-toolbar>=4.2.0
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0